move stateful executor class to thunder/extend
