+++ thunder/extend/__init__.py
@@ -14,6 +14,7 @@ from thunder.core.proxies import Proxy, TensorProxy, proxy
 from thunder.core.pytree import tree_map
 from thunder.core.symbol import Symbol, BoundSymbol, default_python_printer
 from thunder.core.trace import TraceCtx
+from thunder.core.vjp_utils import disable_caching_split_forward_and_backward
 
 __all__ = [
     "register_executor",
@@ -271,6 +272,32 @@ class OperatorExecutor(Executor):
 
         _id = sym_or_id.id if isinstance(sym_or_id, Symbol) else sym_or_id
         self.implmap[_id] = impl
+
+
+class StatefulExecutor(OperatorExecutor):
+    def __init__(self, name, *, version=None):
+        super().__init__(name, version=version)
+        self.state_count: int = 0
+
+    def register_stateful_operator(self, base_name: str, state_class, *, meta):
+        def register_state(*args, **kwargs):
+            state_id = self.state_count
+            name = f"{base_name}_{state_id}"
+
+            def bind_state(bsym):
+                bsym._call_ctx = {name: state_class()}
+
+            sym = self.register_operator(name, meta=meta, bind_postprocess=bind_state)
+
+            self.state_count += 1
+            return sym(*args, *kwargs)
+
+        return register_state
+
+    def get_grad_transform(self, sym: Symbol):
+        grad_transform = super().get_grad_transform(sym)
+        # Always disable cache for stateful grad transform
+        return disable_caching_split_forward_and_backward(grad_transform)
 
 
 class TemporaryExecutor(OperatorExecutor):
